requirements.txt
streamlit
openai
pillow 
numpy